// types/stratzgql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Byte: { input: any; output: any; }
  DateTime: { input: any; output: any; }
  Decimal: { input: any; output: any; }
  Guid: { input: any; output: any; }
  Long: { input: any; output: any; }
  Short: { input: any; output: any; }
  UShort: { input: any; output: any; }
};
// ...
export type CaptainJackIdentityPublicProfileType = {
  __typename?: 'CaptainJackIdentityPublicProfileType';
  captainJackIdentityId?: Maybe<Scalars['Guid']['output']>;
  facebook?: Maybe<Scalars['String']['output']>;
  isAdmin?: Maybe<Scalars['Boolean']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  steamAccount?: Maybe<SteamAccountType>;
  steamAccountId?: Maybe<Scalars['Long']['output']>;
  twitch?: Maybe<Scalars['String']['output']>;
  twitter?: Maybe<Scalars['String']['output']>;
  youTube?: Maybe<Scalars['String']['output']>;
};
// ...
export type FeatType = {
  __typename?: 'FeatType';
  hero?: Maybe<HeroType>;
  heroId?: Maybe<Scalars['Int']['output']>;
  match?: Maybe<MatchType>;
  matchId?: Maybe<Scalars['Long']['output']>;
  type: FeatEnum;
  value?: Maybe<Scalars['Int']['output']>;
};
// ...
export type GuildMemberType = {
  __typename?: 'GuildMemberType';
  guild?: Maybe<GuildType>;
  guildId?: Maybe<Scalars['Int']['output']>;
  imp?: Maybe<Scalars['Int']['output']>;
  joinDateTime?: Maybe<Scalars['Long']['output']>;
  matchCount?: Maybe<Scalars['Int']['output']>;
  steamAccount?: Maybe<SteamAccountType>;
  steamAccountId?: Maybe<Scalars['Int']['output']>;
  winCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type HeroAbilityType = {
  __typename?: 'HeroAbilityType';
  ability?: Maybe<AbilityType>;
  abilityId: Scalars['Short']['output'];
  gameVersionId: Scalars['Short']['output'];
  slot: Scalars['Byte']['output'];
};
// ...
export type HeroDotaPlusLeaderboardRankType = {
  __typename?: 'HeroDotaPlusLeaderboardRankType';
  createdDateTime?: Maybe<Scalars['Long']['output']>;
  heroId?: Maybe<Scalars['Short']['output']>;
  level?: Maybe<Scalars['Byte']['output']>;
  steamAccount?: Maybe<SteamAccountType>;
  steamAccountId?: Maybe<Scalars['Long']['output']>;
  totalActions?: Maybe<Scalars['Long']['output']>;
};

export type HeroDryadType = {
  __typename?: 'HeroDryadType';
  heroId?: Maybe<Scalars['Short']['output']>;
  matchCountVs?: Maybe<Scalars['Long']['output']>;
  matchCountWith?: Maybe<Scalars['Long']['output']>;
  vs?: Maybe<Array<Maybe<HeroStatsHeroDryadType>>>;
  with?: Maybe<Array<Maybe<HeroStatsHeroDryadType>>>;
};

export type HeroFacetType = {
  __typename?: 'HeroFacetType';
  abilityId?: Maybe<Scalars['Short']['output']>;
  facetId?: Maybe<Scalars['Short']['output']>;
  slot?: Maybe<Scalars['Byte']['output']>;
};
// ...
export type HeroLanguageType = {
  __typename?: 'HeroLanguageType';
  displayName?: Maybe<Scalars['String']['output']>;
  hype?: Maybe<Scalars['String']['output']>;
  lore?: Maybe<Scalars['String']['output']>;
};
// ...
export type HeroRoleType = {
  __typename?: 'HeroRoleType';
  level?: Maybe<Scalars['Short']['output']>;
  roleId?: Maybe<HeroRoleEnum>;
};

export type HeroStatType = {
  __typename?: 'HeroStatType';
  agilityBase?: Maybe<Scalars['Float']['output']>;
  agilityGain?: Maybe<Scalars['Float']['output']>;
  attackAcquisitionRange?: Maybe<Scalars['Float']['output']>;
  attackAnimationPoint?: Maybe<Scalars['Float']['output']>;
  attackRange?: Maybe<Scalars['Float']['output']>;
  attackRate?: Maybe<Scalars['Float']['output']>;
  attackType?: Maybe<Scalars['String']['output']>;
  cMEnabled?: Maybe<Scalars['Boolean']['output']>;
  complexity?: Maybe<Scalars['Byte']['output']>;
  enabled?: Maybe<Scalars['Boolean']['output']>;
  heroUnlockOrder?: Maybe<Scalars['Float']['output']>;
  hpBarOffset?: Maybe<Scalars['Float']['output']>;
  hpRegen?: Maybe<Scalars['Float']['output']>;
  intelligenceBase?: Maybe<Scalars['Float']['output']>;
  intelligenceGain?: Maybe<Scalars['Float']['output']>;
  moveSpeed?: Maybe<Scalars['Float']['output']>;
  moveTurnRate?: Maybe<Scalars['Float']['output']>;
  mpRegen?: Maybe<Scalars['Float']['output']>;
  newPlayerEnabled?: Maybe<Scalars['Boolean']['output']>;
  primaryAttribute?: Maybe<Scalars['String']['output']>;
  primaryAttributeEnum?: Maybe<HeroPrimaryAttributeType>;
  startingArmor?: Maybe<Scalars['Float']['output']>;
  startingDamageMax?: Maybe<Scalars['Float']['output']>;
  startingDamageMin?: Maybe<Scalars['Float']['output']>;
  startingMagicArmor?: Maybe<Scalars['Float']['output']>;
  strengthBase?: Maybe<Scalars['Float']['output']>;
  strengthGain?: Maybe<Scalars['Float']['output']>;
  team?: Maybe<Scalars['Boolean']['output']>;
  visionDaytimeRange?: Maybe<Scalars['Float']['output']>;
  visionNighttimeRange?: Maybe<Scalars['Float']['output']>;
};
// ...
export type HeroTalentType = {
  __typename?: 'HeroTalentType';
  abilityId?: Maybe<Scalars['Short']['output']>;
  slot?: Maybe<Scalars['Byte']['output']>;
};

export type HeroType = {
  __typename?: 'HeroType';
  abilities?: Maybe<Array<Maybe<HeroAbilityType>>>;
  aliases?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  displayName?: Maybe<Scalars['String']['output']>;
  facets?: Maybe<Array<Maybe<HeroFacetType>>>;
  gameVersionId?: Maybe<Scalars['Short']['output']>;
  id?: Maybe<Scalars['Short']['output']>;
  language?: Maybe<HeroLanguageType>;
  name?: Maybe<Scalars['String']['output']>;
  roles?: Maybe<Array<Maybe<HeroRoleType>>>;
  shortName?: Maybe<Scalars['String']['output']>;
  stats?: Maybe<HeroStatType>;
  talents?: Maybe<Array<Maybe<HeroTalentType>>>;
};
// ...
export type MatchGroupByType = MatchGroupByAssistsType | MatchGroupByAwardType | MatchGroupByClusterType | MatchGroupByDateDayHeroType | MatchGroupByDateDayType | MatchGroupByDeathsType | MatchGroupByDurationMinutesType | MatchGroupByFactionType | MatchGroupByGameModeType | MatchGroupByGameVersionType | MatchGroupByGoldPerMinuteType | MatchGroupByHeroPerformanceType | MatchGroupByHeroType | MatchGroupByHourType | MatchGroupByIsIntentionalFeedingType | MatchGroupByIsLeagueType | MatchGroupByIsLeaverType | MatchGroupByIsPartyType | MatchGroupByIsRandomType | MatchGroupByIsSeriesType | MatchGroupByIsStatsType | MatchGroupByIsVictoryType | MatchGroupByKillsType | MatchGroupByLaneType | MatchGroupByLeagueIdType | MatchGroupByLevelType | MatchGroupByLobbyTypeType | MatchGroupByPositionType | MatchGroupByRegionType | MatchGroupByRoamLaneType | MatchGroupByRoleType | MatchGroupBySteamAccountIdAgainstTeamType | MatchGroupBySteamAccountIdHeroIdAgainstTeamType | MatchGroupBySteamAccountIdHeroIdType | MatchGroupBySteamAccountIdHeroIdWithTeamType | MatchGroupBySteamAccountIdType | MatchGroupBySteamAccountIdWithTeamType | MatchGroupByTeamType | MatchGroupByTotalKillsType;
// ...
export type PlayerActivitySummaryType = {
  __typename?: 'PlayerActivitySummaryType';
  activity?: Maybe<PlayerBehaviorActivity>;
};

export type PlayerBadgeType = {
  __typename?: 'PlayerBadgeType';
  badgeId?: Maybe<Scalars['Byte']['output']>;
  createdDateTime?: Maybe<Scalars['Long']['output']>;
  slot?: Maybe<Scalars['Byte']['output']>;
};
// ...
export type PlayerHeroPerformanceLongestStreakType = {
  __typename?: 'PlayerHeroPerformanceLongestStreakType';
  currentStreak?: Maybe<Scalars['Int']['output']>;
  heroId?: Maybe<Scalars['Short']['output']>;
  longestStreak?: Maybe<Scalars['Int']['output']>;
};
// ...
export type PlayerHeroesPerformanceType = {
  __typename?: 'PlayerHeroesPerformanceType';
  avgAssists?: Maybe<Scalars['Float']['output']>;
  avgDeaths?: Maybe<Scalars['Float']['output']>;
  avgKills?: Maybe<Scalars['Float']['output']>;
  best?: Maybe<Scalars['Float']['output']>;
  duration: Scalars['Int']['output'];
  experiencePerMinute: Scalars['Int']['output'];
  goldPerMinute: Scalars['Int']['output'];
  hero?: Maybe<HeroType>;
  heroId: Scalars['Short']['output'];
  imp?: Maybe<Scalars['Int']['output']>;
  kDA?: Maybe<Scalars['Float']['output']>;
  lastPlayedDateTime?: Maybe<Scalars['Long']['output']>;
  matchCount: Scalars['Int']['output'];
  positionScore?: Maybe<Array<Maybe<PlayerHeroesPerformanceScoreType>>>;
  winCount: Scalars['Int']['output'];
};
// ...
export type PlayerPerformanceType = {
  __typename?: 'PlayerPerformanceType';
  assists?: Maybe<Scalars['Int']['output']>;
  assistsAverage?: Maybe<Scalars['Decimal']['output']>;
  awardMatchCount: Scalars['Int']['output'];
  composition?: Maybe<PlayerPerformanceCompositionType>;
  cs?: Maybe<Scalars['Int']['output']>;
  csAverage?: Maybe<Scalars['Decimal']['output']>;
  deaths?: Maybe<Scalars['Int']['output']>;
  deathsAverage?: Maybe<Scalars['Decimal']['output']>;
  gpm?: Maybe<Scalars['Int']['output']>;
  gpmAverage?: Maybe<Scalars['Decimal']['output']>;
  heroId?: Maybe<Array<Maybe<Scalars['Short']['output']>>>;
  imp?: Maybe<Scalars['Int']['output']>;
  kills?: Maybe<Scalars['Int']['output']>;
  killsAverage?: Maybe<Scalars['Decimal']['output']>;
  matchCount: Scalars['Int']['output'];
  maxStreak: Scalars['Int']['output'];
  mmrBracket: Scalars['Int']['output'];
  mmrTier: Scalars['Int']['output'];
  mvpCount: Scalars['Int']['output'];
  /** Contains an array of 6 items which are listed as index 0 - 1st Pick, Index 1 - Pick 2nd, 3rd, Index 2 - Pick 4th, 5th, Index 3 - Pick 6th, Pick 7th, Index 4 - Pick 8th, 9ths, Index 5 - Pick 10th */
  pickOrder?: Maybe<Array<Maybe<Scalars['Int']['output']>>>;
  position?: Maybe<Array<Maybe<PlayerPerformancePositionType>>>;
  rank?: Maybe<Scalars['Int']['output']>;
  streak: Scalars['Int']['output'];
  topCoreCount: Scalars['Int']['output'];
  topSupportCount: Scalars['Int']['output'];
  winCount: Scalars['Int']['output'];
  xpm?: Maybe<Scalars['Int']['output']>;
  xpmAverage?: Maybe<Scalars['Decimal']['output']>;
};
// ...
export type PlayerType = {
  __typename?: 'PlayerType';
  activity?: Maybe<PlayerActivitySummaryType>;
  badges?: Maybe<Array<Maybe<PlayerBadgeType>>>;
  behaviorScore?: Maybe<Scalars['Short']['output']>;
  /** Gets the players of Dota which have DotaPlus and have a high level hero. */
  dotaPlus?: Maybe<Array<Maybe<HeroDotaPlusLeaderboardRankType>>>;
  /** A list of the high achivement skills by a Player. */
  feats?: Maybe<Array<Maybe<FeatType>>>;
  firstMatchDate?: Maybe<Scalars['Long']['output']>;
  guildMember?: Maybe<GuildMemberType>;
  heroPerformance?: Maybe<PlayerPerformanceType>;
  /** A list of the current Streak and the Longest Streak for each Hero by a Player. */
  heroStreaks?: Maybe<Array<Maybe<PlayerHeroPerformanceLongestStreakType>>>;
  /** Returns a list of all heroes played by the steam account id and contains data about the average performance. */
  heroesPerformance?: Maybe<Array<Maybe<PlayerHeroesPerformanceType>>>;
  identity?: Maybe<CaptainJackIdentityPublicProfileType>;
  imp?: Maybe<Scalars['Int']['output']>;
  isFollowed?: Maybe<Scalars['Boolean']['output']>;
  lastMatchDate?: Maybe<Scalars['Long']['output']>;
  lastMatchRegionId?: Maybe<Scalars['Byte']['output']>;
  leaderboardRanks?: Maybe<Array<Maybe<SteamAccountSeasonLeaderBoardRankType>>>;
  matchCount?: Maybe<Scalars['Int']['output']>;
  /** Find match details by steam account id. steamAccountId is a required input field. */
  matches?: Maybe<Array<Maybe<MatchType>>>;
  /** Find match details by steam account id. The return is modified to group the data by the GroupBy parameter. */
  matchesGroupBy?: Maybe<Array<Maybe<MatchGroupByType>>>;
  names?: Maybe<Array<Maybe<SteamAccountNameType>>>;
  performance?: Maybe<PlayerPerformanceType>;
  ranks?: Maybe<Array<Maybe<SteamAccountSeasonRankType>>>;
  simpleSummary?: Maybe<PlayerCardHoverType>;
  steamAccount?: Maybe<SteamAccountType>;
  steamAccountId?: Maybe<Scalars['Long']['output']>;
  team?: Maybe<SteamAccountTeamMemberType>;
  winCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type SteamAccountSeasonLeaderBoardRankType = {
  __typename?: 'SteamAccountSeasonLeaderBoardRankType';
  asOfDateTime?: Maybe<Scalars['Long']['output']>;
  rank?: Maybe<Scalars['Short']['output']>;
  seasonLeaderBoardDivisionId?: Maybe<LeaderboardDivision>;
  seasonRankId?: Maybe<Scalars['Byte']['output']>;
  steamAccountId?: Maybe<Scalars['Long']['output']>;
};